moved process owner to database service
